proper handling of the channel in EventSelection
Here, I introduce a drastic change to the way muChannel is handled by the event selection and the corresponding control plot.
For Z candidates (and correspondingly good jets), we now always use _all, with a check of the candidate kind in the category check.
This should simplify everything and make the code compliant with the new generalized ControlPlot.py
It also avoid double counting between ee and mm categories for eemmbb events... admitedly rare.

Such a change will have to be propagated to other control plots.