Moved the zbbCommons items to zbbConfig
Rationalization of the configuration infrastructure.
All classes now rely solely on zbbConfig. The use of zbbCommons is deprecated.
For compatibility reasons, a zbbCommons file has been added both in analysisScripts and pyrootScripts.
Nevertheless, none of these files is a full copy of the original. It contains only labels for
legacy analysis scripts, and norm factors for pyroot scripts.
For the classes in the base directory, there is no change of behavior,
except for reweighting classes, where no default input file is now given.