Reworked zbb_counter to make use of the new infrastructure
This is much more compact, as most of the work is done elsewhere.
The code to reconstruct the gen-level Z candidate has also been moved to MonteCarloSelection. It includes a version starting from the Gen Z and another one starting from the gen leptons. Results are slightly different between the two versions.