Rework of the Efficiency_acceptance to benefit from the PatAnalysis.
This code has been reengineered after the move to PatAnalysis.
It is running, and should give an output close to the previous version,
but the focus has not been to be totally equivalent.
It serves mostly as a example as it is now and it should be checked that
the code does actually what you intend it to do.