Deprecated globalLists and zbbCommons everywhere
Now uses the database interface from zbbSamples.py to get lumi and path to samples.
Various sample lists are also grouped in zbbSamples.py.
Warning: the change is just formal and was not tested. If a script has to be used, then it should be reviewed in detail first.
Especially, the naming conventions should be checked, and the dependancy on these in the code minimized.

In zbbSamples.py, added a set of utility functions to ease interaction with db and sample lists.
While probably not fully functional, the present picture is now coherent and shows how SAMADhi integrates with the analysis scripts without the need for a duplication of information in python files.